# web/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Machina Host (Daemon) API",
    "version": "1.0.0",
    "description": "Hypervisor host REST API (machina-daemon): libvirt VMs, storage, networks, K8s proxies, backups, and host tools."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same origin as Machina UI"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "cookieAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer token (mach_… API key or session JWT)"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "machina_session",
        "description": "Browser session cookie"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/platform/controller/{*rest}": {
      "delete": {
        "summary": "DELETE /api/v1/platform/controller/{*rest}",
        "tags": [
          "platform"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "500": {
            "description": "Server error"
          }
        },
        "parameters": [
          {
            "name": "*rest",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "get": {
        "summary": "GET /api/v1/platform/controller/{*rest}",
        "tags": [
          "platform"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "500": {
            "description": "Server error"
          }
        },
        "parameters": [
          {
            "name": "*rest",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "patch": {
        "summary": "PATCH /api/v1/platform/controller/{*rest}",
        "tags": [
          "platform"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "500": {
            "description": "Server error"
          }
        },
        "parameters": [
          {
            "name": "*rest",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      },
      "post": {
        "summary": "POST /api/v1/platform/controller/{*rest}",
        "tags": [
          "platform"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "500": {
            "description": "Server error"
          }
        },
        "parameters": [
          {
            "name": "*rest",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      },
      "put": {
        "summary": "PUT /api/v1/platform/controller/{*rest}",
        "tags": [
          "platform"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "500": {
            "description": "Server error"
          }
        },
        "parameters": [
          {
            "name": "*rest",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
...
    "/api/v1/system/auth/ldap-settings": {
      "get": {
        "summary": "GET /api/v1/system/auth/ldap-settings",
        "tags": [
          "system"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/v1/system/auth/ldap-test": {
      "post": {
        "summary": "POST /api/v1/system/auth/ldap-test",
        "tags": [
          "system"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "500": {
            "description": "Server error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
    "/api/v1/system/auth/oidc-settings": {
      "get": {
        "summary": "GET /api/v1/system/auth/oidc-settings",
        "tags": [
          "system"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/v1/system/auth/saml-settings": {
      "get": {
        "summary": "GET /api/v1/system/auth/saml-settings",
        "tags": [
          "system"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/v1/system/create-vm-defaults": {
      "get": {
        "summary": "GET /api/v1/system/create-vm-defaults",
        "tags": [
          "system"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/v1/system/observability-settings": {
      "get": {
        "summary": "GET /api/v1/system/observability-settings",
        "tags": [
          "system"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/v1/system/os-users": {
      "post": {
        "summary": "POST /api/v1/system/os-users",
        "tags": [
          "system"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "500": {
            "description": "Server error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
    "/api/v1/system/os-users/{username}": {
      "delete": {
        "summary": "DELETE /api/v1/system/os-users/{username}",
        "tags": [
          "system"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "500": {
            "description": "Server error"
          }
        },
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v1/system/os-users/capability": {
      "get": {
        "summary": "GET /api/v1/system/os-users/capability",
        "tags": [
          "system"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/v1/system/platform-info": {
      "get": {
        "summary": "GET /api/v1/system/platform-info",
        "tags": [
          "system"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  }
}